changed VM size, starting porting of old code

+++ src/utils/printf.c
@@ -85,8 +85,8 @@ static void* stdout_putp;
 
 static void uli2a(unsigned long int num, unsigned int base, int uc,char * bf)
 	{
-	int n=0;
-	unsigned int d=1;
+	long n=0;
+	unsigned long d=1;
 	while (num/d >= base)
 		d*=base;
 	while (d!=0) {
